ADD reusable function of array-queue for empty array index.
AS IS : when front and rear of array-queue are same as increasing continuously in the end, it cannot reuse previous array-buffer-index area empty.
TO BE : array-queue reuses empty array index previous by using memcpy to copy the front index of array the data to enqueue and initialize front and rear again at the above situation.